Fix download path for resnext and wideresnet models
The code used to call wget with a command line that made it
download the .pth-files into the current working directory instead of
the model path, causing the .pth-files to not be found. The '-O'
parameter was added to wget to make it download the files to the correct
location.